:recycle: move submission folder
